[optimization] timer optimization

+++ SourceCode/HAL/include/hal/el.h
@@ -1,7 +1,9 @@
 #pragma once
 
+#include <stdint.h>
+
 /**
  * get current exception level
  * @return current exception level
  */
-int get_current_el();
+uint64_t get_current_el();
